Fixing the bug when you run game for first time
When we run the game for first time, We read game level from file. If
file doesn't exist, it was giving error previously. Now fixing this
error by checking whether file exists or not.
Also added a global variable to track game level which will be needed by
other components.