Bug fixes
Fixed a bug in leader board to insert score properly. The files will
always have scores in ascending order of ranks. Ranks become implicit
now.

Enhanced timer feature to be similar to real game 0 to 999.

Fixed a bug in timer that only after first click the timer should start
and not before that.